requirements.txt
# YouTube handling
yt-dlp>=2023.12.0
youtube-transcript-api>=0.6.0

# AI APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Database
# sqlite3 is built-in

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.66.0

# Optional - for data export
pandas>=2.0.0

# Web UI
Flask>=3.0.0

# Optional - for testing
pytest>=7.4.0
